+++ packages/app-content-pages/pages/team.js
@@ -1,6 +1,5 @@
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import createTeamResponse from '../src/api/team/team.js'
-import logNodeError from '../src/helpers/logger/logNodeError.js'
 export { default } from '../src/screens/Teams'
 
 export async function getStaticProps({ locale }) {
@@ -27,7 +26,6 @@ export async function getStaticProps({ locale }) {
       revalidate: 60 * 60 * 1 // 1 hour
     }
   } catch (error) {
-    logNodeError(error)
     throw error
   }
 }