+++ packages/app-content-pages/src/helpers/logger/index.js
@@ -1,2 +1,0 @@
-export { default as logNodeError } from './logNodeError'
-export { default as logReactError } from './logReactError'