+++ packages/app-content-pages/src/screens/Teams/components/Team/Team.js
@@ -9,11 +9,12 @@ function Team({ name = '', people = [], slug = '' }) {
     <Box as='section' key={name}>
       <HeadingForNav
         id={slug}
-        color='black'
-        level='2'
+        color={{ light: 'black', dark: 'white' }}
+        level={2}
         size='1.5rem'
         tabIndex={-1}
         textAlign='center'
+        style={{ padding: '30px 0' }}
       >
         {name}
       </HeadingForNav>