+++ packages/app-project/src/screens/ProjectHomePage/components/ZooniverseTalk/components/RecentSubjects/RecentSubjectsContainer.js
@@ -1,4 +1,3 @@
-import asyncStates from '@zooniverse/async-states'
 import { MobXProviderContext, observer } from 'mobx-react'
 import { bool } from 'prop-types'
 import { useContext } from 'react'