+++ packages/app-root/src/components/PageHeader.js
@@ -4,13 +4,15 @@ import { useUnreadMessages, useUnreadNotifications } from '@zooniverse/react-com
 import auth from 'panoptes-client/lib/auth'
 import { useContext, useState } from 'react'
 
-import { PanoptesAuthContext } from '../contexts'
+import { PanoptesAuthContext, ThemeModeContext } from '../contexts'
 
 export default function PageHeader() {
   const { adminMode, user } = useContext(PanoptesAuthContext)
   const { data: unreadMessages }= useUnreadMessages(user)
   const { data: unreadNotifications }= useUnreadNotifications(user)
   const [activeIndex, setActiveIndex] = useState(-1)
+
+  const { themeMode, toggleTheme } = useContext(ThemeModeContext)
 
   function openRegisterModal() {
     setActiveIndex(1)
@@ -36,10 +38,13 @@ export default function PageHeader() {
         onActive={setActiveIndex}
       />
       <ZooHeader
-        isAdmin={adminMode}
+        adminMode={adminMode}
+        onThemeChange={toggleTheme}
         register={openRegisterModal}
+        showThemeToggle
         signIn={openSignInModal}
         signOut={onSignOut}
+        themeMode={themeMode}
         unreadMessages={unreadMessages}
         unreadNotifications={unreadNotifications}
         user={user}