+++ packages/lib-react-components/src/ZooHeader/ZooHeader.js
@@ -11,6 +11,7 @@ import MainNavList from './components/MainNavList'
 import NarrowMainNavMenu from './components/NarrowMainNavMenu'
 import UserNavigation from './components/UserNavigation/UserNavigation.js'
 import ZooniverseLogo from '../ZooniverseLogo'
+import ThemeModeToggle from './components/ThemeModeToggle/ThemeModeToggle'
 
 export const StyledHeader = styled(Box)`
   color: #b2b2b2;
@@ -37,11 +38,14 @@ const defaultHandler = () => true
 
 export default function ZooHeader({
   breakpoint = 960,
-  isAdmin = false,
+  adminMode = false,
   isNarrow = false,
+  onThemeChange = defaultHandler,
   register = defaultHandler,
+  showThemeToggle = false,
   signIn = defaultHandler,
   signOut = defaultHandler,
+  themeMode = 'light',
   unreadMessages = 0,
   unreadNotifications = 0,
   user = {},
@@ -97,7 +101,7 @@ export default function ZooHeader({
         <MainNavList
           adminNavLinkLabel={adminNavLinkLabel}
           adminNavLinkURL={adminNavLinkURL}
-          isAdmin={user?.admin && isAdmin}
+          adminMode={user?.admin && adminMode}
           isNarrow={isNarrow}
           mainHeaderNavListLabels={mainHeaderNavListLabels}
           mainHeaderNavListURLs={mainHeaderNavListURLs}
@@ -111,9 +115,12 @@ export default function ZooHeader({
         {hasMounted && (
           <UserNavigation
             isNarrow={isNarrow}
+            onThemeChange={onThemeChange}
             register={register}
+            showThemeToggle={showThemeToggle}
             signIn={signIn}
             signOut={signOut}
+            themeMode={themeMode}
             unreadMessages={unreadMessages}
             unreadNotifications={unreadNotifications}
             user={user}
@@ -123,7 +130,7 @@ export default function ZooHeader({
           <NarrowMainNavMenu
             adminNavLinkLabel={adminNavLinkLabel}
             adminNavLinkURL={adminNavLinkURL}
-            isAdmin={user?.admin && isAdmin}
+            adminMode={user?.admin && adminMode}
             mainHeaderNavListLabels={mainHeaderNavListLabels}
             mainHeaderNavListURLs={mainHeaderNavListURLs}
           />
@@ -134,15 +141,18 @@ export default function ZooHeader({
 }
 
 ZooHeader.propTypes = {
-  isAdmin: PropTypes.bool,
+  adminMode: PropTypes.bool,
   isNarrow: PropTypes.bool,
+  onThemeChange: PropTypes.func,
   register: PropTypes.func,
+  showThemeToggle: PropTypes.bool,
   signIn: PropTypes.func.isRequired,
   signOut: PropTypes.func.isRequired,
+  themeMode: PropTypes.string,
   unreadMessages: PropTypes.number,
   unreadNotifications: PropTypes.number,
   user: PropTypes.shape({
-    display_name: PropTypes.string.isRequired,
-    login: PropTypes.string.isRequired
+    display_name: PropTypes.string,
+    login: PropTypes.string
   })
 }