+++ packages/lib-react-components/src/ZooHeader/ZooHeader.stories.js
@@ -1,3 +1,4 @@
+import { useState } from 'react'
 import ZooHeader from './ZooHeader'
 import readme from './README.md'
 
@@ -21,6 +22,48 @@ export default {
   }
 }
 
+export function ThemeToggleSignedIn({ signIn, signOut }) {
+  const [themeMode, setThemeMode] = useState('light')
+  function onThemeChange() {
+    const newTheme = (themeMode === 'light') ? 'dark' : 'light'
+    setThemeMode(newTheme)
+  }
+
+  return (
+    <ZooHeader
+      onThemeChange={onThemeChange}
+      showThemeToggle
+      signIn={signIn}
+      signOut={signOut}
+      themeMode={themeMode}
+      user={{
+        display_name: 'zootester1',
+        login: 'zootester1'
+      }}
+    />
+  )
+}
+
+export function ThemeToggleSignedOut({ signIn, signOut }) {
+  const [themeMode, setThemeMode] = useState('light')
+  function onThemeChange() {
+    const newTheme = (themeMode === 'light') ? 'dark' : 'light'
+    setThemeMode(newTheme)
+  }
+
+  return (
+    <ZooHeader
+      onThemeChange={onThemeChange}
+      showThemeToggle
+      signIn={signIn}
+      signOut={signOut}
+      themeMode={themeMode}
+      user={{}}
+    />
+  )
+}
+
+
 export function SignedOut({ signIn, signOut }) {
   return <ZooHeader signIn={signIn} signOut={signOut} user={{}} />
 }
@@ -40,10 +83,10 @@ export function SignedIn({ signIn, signOut }) {
 }
 
 
-export function SignedInAsAdmin({ signIn, signOut }) {
+export function SignedInAdminMode({ signIn, signOut }) {
   return (
     <ZooHeader
-      isAdmin
+      adminMode
       signIn={signIn}
       signOut={signOut}
       user={{